refactor: thin apps-list route and split imageGen setup into its own module (#2533)
Move the GET /api/apps handler's inline PM2 enrichment into a new
appListEnrichment service (enrichAppsWithPm2Status), so the route validates
and responds while the service owns the per-pm2Home reads, absent-vs-empty
status classification, ecosystem-config process backfill, and port derivation.
The pure port/status helpers (deriveUiPort/deriveAppPorts/computeOverallStatus)
relocate to the service and are re-exported from routes/apps/shared.js so the
detail/lifecycle routes keep one shared implementation.

Split the /api/image-gen/setup/* handlers (python probe/venv bootstrap, FLUX.2
install/status, HuggingFace-token store, pip installer) out of the large
imageGen render router into routes/imageGenSetup.js, mounted at /setup. The
setup-check cache and pip allowlist move to services/imageGen/setup.js.

Pure mechanical refactor; existing route suites cover both paths.
