+++ server/routes/imageGen.js
@@ -11,7 +11,7 @@
 
 import { Router } from 'express';
 import { z } from 'zod';
-import { unlink, stat, copyFile } from 'fs/promises';
+import { unlink, copyFile } from 'fs/promises';
 import { asyncHandler, ServerError, failValidation } from '../lib/errorHandler.js';
 import {
   validateRequest, imageEdgeSchema, refineImagePixelCap, PIXEL_CAP_MESSAGE,
@@ -19,20 +19,12 @@ import {
 import { optionalUploadFields } from '../lib/multipart.js';
 import * as imageGen from '../services/imageGen/index.js';
 import { local, IMAGE_GEN_MODE, IMAGE_GEN_MODES } from '../services/imageGen/index.js';
+import setupRouter from './imageGenSetup.js';
 import { enqueueJob, attachSseClient as attachQueueSseClient, cancelJob, listJobs } from '../services/mediaJobQueue/index.js';
-import { getSettings, updateSettingsWith } from '../services/settings.js';
-import { getHfToken, getHfTokenInfo, HF_TOKEN_REGEX } from '../lib/hfToken.js';
 import { getImageModels, isFlux2, isEditOnly, repoForModel, requiredReposForModel } from '../lib/mediaModels.js';
 import { usesDiffusersRunner } from '../lib/runners.js';
 import { inspectModelCache, verifyModelCache, repairModelCache, aggregateVerifies } from '../lib/hfCache.js';
-import { startHfDownloadStream, openSseStream } from '../lib/sseDownload.js';
-import { createInstallLogger } from '../lib/installLogger.js';
-import {
-  REQUIRED_PACKAGES, detectPython, installPackages,
-  createVenv, isAllowedPython, pipNameFor,
-  resolveFlux2Python, FLUX2_VENV_DEFAULT, installFlux2Venv, isFlux2VenvHealthy,
-  detectArm64Python, HOST_ARCH, probePythonHealth,
-} from '../lib/pythonSetup.js';
+import { startHfDownloadStream } from '../lib/sseDownload.js';
 import { PATHS, ensureDir, resolveGalleryImage } from '../lib/fileUtils.js';
 import { prepareGenerateParams } from '../services/imageGen/prepareParams.js';
 import { applyImageClean, applyWatermarkRemoval, applyLightRegenVariant } from '../services/imageGen/variants.js';
@@ -799,300 +791,8 @@ router.post('/:filename/regenerate', asyncHandler(async (req, res) => {
   return res.json(queuedImageResponse({ ...queued, mode: IMAGE_GEN_MODE.LOCAL, model: backend.model.id }));
 }));
 
-// --- Local-mode setup automation ---
-
-router.get('/setup/python', asyncHandler(async (_req, res) => {
-  const path = await detectPython();
-  res.json({ path });
-}));
-
-// SSE-driven FLUX.2 venv bootstrap. Replaces the "drop to a shell and run
-// INSTALL_FLUX2=1 bash scripts/setup-image-video.sh" friction with an in-app
-// install: the client opens an EventSource, gets staged progress events
-// (detect → venv → upgrade-pip → install → verify), and either finishes or
-// surfaces a clear error. Runs the install logic in-process via
-// installFlux2Venv() so we get structured `stage` events the UI can animate
-// against, instead of having to parse bash output.
-//
-// In-flight singleton: a rapid double-click would otherwise race two pip
-// processes against the same venv directory. resolveFlux2Python() can't
-// gate the second click — the first install hasn't created the python yet.
-let flux2InstallInFlight = null;
-
-router.get('/setup/flux2-install', asyncHandler(async (req, res) => {
-  const { send, safeEnd } = openSseStream(res);
-
-  // Skip only when the venv binary AND the import work — a half-broken venv
-  // (binary present, packages missing from a killed mid-install) needs to
-  // re-run the install, not be reported as ready.
-  if (await isFlux2VenvHealthy()) {
-    send({ type: 'stage', stage: 'verify', message: 'FLUX.2 venv already installed.' });
-    send({ type: 'complete', message: 'Already installed — nothing to do.' });
-    return safeEnd();
-  }
-  if (flux2InstallInFlight) {
-    send({ type: 'error', message: 'Another FLUX.2 install is already running. Wait for it to finish or restart PortOS.' });
-    return safeEnd();
-  }
-
-  // Server-console visibility for the multi-GB torch install (start / stage
-  // milestones / outcome) — installFlux2Venv streams progress only to `send`.
-  const installLog = createInstallLogger({ installer: 'FLUX.2 venv', target: FLUX2_VENV_DEFAULT });
-  const emit = (ev) => { installLog.onEvent(ev); send(ev); };
-  installLog.start();
-
-  const { promise, kill } = installFlux2Venv(emit);
-  flux2InstallInFlight = promise;
-  promise
-    // installFlux2Venv resolves { ok:false } on some pip failures without
-    // emitting a terminal SSE event, so reconcile the outcome from the result
-    // (a no-op if `onEvent` already logged a complete/error frame).
-    .then((result) => {
-      if (result?.ok) installLog.success(result?.pythonPath ? `ready: ${result.pythonPath}` : undefined);
-      else installLog.failure(`failed at stage ${result?.stage || 'unknown'}`);
-    })
-    .catch((err) => emit({ type: 'error', message: err?.message || 'Unknown installer failure' }))
-    .finally(() => {
-      flux2InstallInFlight = null;
-      safeEnd();
-    });
-
-  // Cancel the install if the client navigates away mid-bootstrap. A torch
-  // install is a multi-GB download and would otherwise keep running invisibly.
-  req.on('close', () => { installLog.cancel(); kill(); safeEnd(); });
-}));
-
-// Used by the FLUX.2 model picker: surface a banner when the gated repo's
-// license hasn't been accepted (HF_TOKEN missing) and the runner is set up.
-// `venvInstalled` reflects functional health (binary AND packages import) —
-// a half-broken venv would otherwise hide the install banner forever.
-router.get('/setup/flux2-status', asyncHandler(async (req, res) => {
-  const [token, healthy] = await Promise.all([getHfToken(), isFlux2VenvHealthy()]);
-  const venvPython = resolveFlux2Python();
-  // The 9B (bf16) and 4B variants ship as separately-gated repos with
-  // distinct HF license URLs. Use the active model's `licenseUrl` when the
-  // client supplies a `modelId`; fall back to the 4B URL for callers that
-  // pre-date the multi-variant registry.
-  const FLUX2_DEFAULT_LICENSE = 'https://huggingface.co/black-forest-labs/FLUX.2-klein-4B';
-  let licenseUrl = FLUX2_DEFAULT_LICENSE;
-  if (typeof req.query?.modelId === 'string' && req.query.modelId.length > 0) {
-    const model = getImageModels().find((m) => m.id === req.query.modelId);
-    if (isFlux2(model) && typeof model?.licenseUrl === 'string' && model.licenseUrl.length > 0) {
-      licenseUrl = model.licenseUrl;
-    }
-  }
-  res.json({
-    hfTokenPresent: !!token,
-    venvInstalled: healthy,
-    venvPath: venvPython,
-    expectedVenvPath: FLUX2_VENV_DEFAULT,
-    licenseUrl,
-  });
-}));
-
-// Generic HF-token presence check for legacy mflux runners that don't need
-// the FLUX.2 venv. Any model entry with `requiresHfToken: true` in
-// data/media-models.json drives the banner through this endpoint.
-router.get('/setup/hf-token-status', asyncHandler(async (_req, res) => {
-  const { token, source } = await getHfTokenInfo();
-  res.json({ hfTokenPresent: !!token, source });
-}));
-
-// Save the HF token from the inline form on the Image Gen page. settings.json
-// is the canonical location (single-user app behind Tailscale — see CLAUDE.md).
-// Same endpoint serves FLUX.2 and legacy mflux gated models — the token is
-// global (HF_TOKEN env in spawn).
-const hfTokenSchema = z.object({
-  token: z.string().regex(HF_TOKEN_REGEX, 'Token must look like `hf_…`').max(200),
-});
-router.post('/setup/hf-token', asyncHandler(async (req, res) => {
-  const { token } = validateRequest(hfTokenSchema, req.body || {});
-  await updateSettingsWith((settings) => ({
-    ...settings,
-    imageGen: { ...(settings.imageGen || {}), hfToken: token.trim() },
-  }));
-  res.json({ ok: true, hfTokenPresent: true, source: 'stored' });
-}));
-
-// Clear the stored HF token. Falls back to env / CLI tokens if present —
-// callers should re-fetch /setup/hf-token-status to see the post-clear state.
-router.delete('/setup/hf-token', asyncHandler(async (_req, res) => {
-  await updateSettingsWith((settings) => {
-    const { hfToken: _drop, ...restImageGen } = settings.imageGen || {};
-    return { ...settings, imageGen: restImageGen };
-  });
-  const { token, source } = await getHfTokenInfo();
-  res.json({ ok: true, hfTokenPresent: !!token, source });
-}));
-
-const checkSchema = z.object({ pythonPath: z.string().min(1) });
-
-// /setup/check is called on every keystroke in the python-path input (debounced
-// to 400ms), on mount, AND on the refresh button — each call spawns a python
-// subprocess (~0.5-1s warm) plus the optional `detectArm64Python` walk. A
-// modest in-memory cache, keyed by (pythonPath, stat.mtimeMs) and bounded by
-// SETUP_CHECK_TTL_MS, collapses the typing-flow repeats to memo hits without
-// risking stale data — the key changes the moment the interpreter is swapped
-// (venv create, brew upgrade) and the install path explicitly busts on
-// completion.
-const SETUP_CHECK_TTL_MS = 30_000;
-const setupCheckCache = new Map();
-
-const buildSetupCheck = async (pythonPath) => {
-  const health = await probePythonHealth(pythonPath);
-  // The arch warning is specifically about mlx wheels (arm64-only) on Apple
-  // Silicon. A generic interpreterArch !== HOST_ARCH compare would false-
-  // positive on Windows (Python reports `AMD64`, Node reports `x86_64`) and
-  // on hypothetical arm64 Linux — where mlx isn't even in REQUIRED_PACKAGES.
-  const archMismatch = process.platform === 'darwin'
-    && HOST_ARCH === 'arm64'
-    && health.interpreterArch === 'x86_64';
-  const suggestedArm64Python = archMismatch ? await detectArm64Python() : null;
-  return {
-    pythonPath,
-    required: REQUIRED_PACKAGES,
-    hostArch: HOST_ARCH,
-    archMismatch,
-    suggestedArm64Python,
-    ...health,
-  };
-};
-
-const invalidateSetupCheck = (pythonPath) => {
-  if (!pythonPath) {
-    setupCheckCache.clear();
-    return;
-  }
-  const prefix = `${pythonPath}|`;
-  for (const key of setupCheckCache.keys()) {
-    if (key.startsWith(prefix)) setupCheckCache.delete(key);
-  }
-};
-
-router.get('/setup/check', asyncHandler(async (req, res) => {
-  const { pythonPath } = validateRequest(checkSchema, req.query);
-  if (!isAllowedPython(pythonPath)) {
-    throw new ServerError('pythonPath must be a python interpreter (basename python/python3/python3.NN)', { status: 400 });
-  }
-  // mtime keys auto-bust when the interpreter binary itself changes (rare but
-  // surfaces brew upgrades / re-symlinks). A stat() failure (path not found)
-  // skips the cache rather than poisoning it with a `mtime=missing` entry.
-  const mtimeMs = await stat(pythonPath).then((s) => s.mtimeMs).catch(() => null);
-  const key = mtimeMs !== null ? `${pythonPath}|${mtimeMs}` : null;
-  if (key) {
-    const hit = setupCheckCache.get(key);
-    if (hit && hit.expiresAt > Date.now()) return res.json(hit.result);
-  }
-  const result = await buildSetupCheck(pythonPath);
-  if (key) {
-    // Sweep expired entries on every write so a long-running process doesn't
-    // accumulate stale (path, mtime) combos from intermediate keystrokes
-    // (each typed character of a path string lands a unique key here).
-    const now = Date.now();
-    for (const [k, v] of setupCheckCache) {
-      if (v.expiresAt <= now) setupCheckCache.delete(k);
-    }
-    setupCheckCache.set(key, { result, expiresAt: now + SETUP_CHECK_TTL_MS });
-  }
-  res.json(result);
-}));
-
-const venvSchema = z.object({
-  basePython: z.string().min(1).optional(),
-});
-
-router.post('/setup/create-venv', asyncHandler(async (req, res) => {
-  const { basePython } = validateRequest(venvSchema, req.body || {});
-  if (basePython && !isAllowedPython(basePython)) {
-    throw new ServerError('basePython must be a python interpreter (basename python/python3/python3.NN)', { status: 400 });
-  }
-  const base = basePython || (await detectPython());
-  if (!base) {
-    throw new ServerError('No base Python 3 found to bootstrap a venv. Install Python 3.10+ first.', { status: 400 });
-  }
-  const target = join(PATHS.data, 'python', 'venv');
-  const venvPython = await createVenv(base, target);
-  // Bust the setup-check cache for both the base interpreter and the new
-  // venv python — the venv inherits the base's mtime-key but its packages
-  // differ, and a subsequent /setup/check would otherwise return the base's
-  // pre-venv snapshot.
-  invalidateSetupCheck(base);
-  invalidateSetupCheck(venvPython);
-  res.json({ pythonPath: venvPython, target });
-}));
-
-// Allowlist: only PortOS's own required pip names (or their pinned variants
-// like `transformers<5`) are installable. Without this, the endpoint would
-// happily pip-install arbitrary PyPI packages — the install runs as the
-// PortOS user and pip itself executes setup.py from the package, so an
-// arbitrary package install is effectively arbitrary code execution.
-// Build the pip-spec allowlist from REQUIRED_PACKAGES via pipNameFor — that
-// translates import names (`cv2`) to their actual pip specs
-// (`opencv-python`). Without this mapping, the allowlist would contain
-// import-only names that can't actually be installed but ALSO don't appear
-// here as their pip specs, so the legitimate install request would 400.
-// Worse: an import name like `cv2` isn't a real PyPI package but if a
-// typosquat existed under that name it'd be installable.
-const REQUIRED_PIP_NAMES = new Set([
-  ...REQUIRED_PACKAGES.map(pipNameFor),
-  // Windows torch path also installs torch + diffusers, which are in
-  // REQUIRED_PACKAGES on Windows but not on macOS — keep them allowlisted
-  // unconditionally so a Windows install requested from a macOS server
-  // (unlikely but possible) doesn't 400 unhelpfully.
-  'torch',
-  'diffusers',
-  // Both the bare `transformers` and the macOS-pinned `transformers<5`
-  // variant should be installable; pipNameFor only emits the pinned
-  // variant on macOS, so list both unconditionally for safety.
-  'transformers',
-  'transformers<5',
-]);
-
-const installSchema = z.object({
-  pythonPath: z.string().min(1),
-  packages: z.array(z.string().min(1)).min(1).max(40),
-});
-
-// EventSource consumers re-run /setup/check on `complete` to refresh status.
-router.get('/setup/install', (req, res) => {
-  const pythonPath = req.query.pythonPath;
-  const packages = String(req.query.packages || '').split(',').filter(Boolean);
-  const parsed = installSchema.safeParse({ pythonPath, packages });
-  if (!parsed.success) {
-    res.writeHead(400, { 'Content-Type': 'application/json' });
-    return res.end(JSON.stringify({ error: parsed.error.message }));
-  }
-  if (!isAllowedPython(parsed.data.pythonPath)) {
-    res.writeHead(400, { 'Content-Type': 'application/json' });
-    return res.end(JSON.stringify({ error: 'pythonPath must be a python interpreter' }));
-  }
-  const disallowed = parsed.data.packages.filter((p) => !REQUIRED_PIP_NAMES.has(p));
-  if (disallowed.length) {
-    res.writeHead(400, { 'Content-Type': 'application/json' });
-    return res.end(JSON.stringify({ error: `Packages not in allowlist: ${disallowed.join(', ')}` }));
-  }
-
-  // `send` and `safeEnd` from openSseStream no-op once the response has ended
-  // so a late pip-output line (or the promise.then below) doesn't trigger
-  // ERR_STREAM_WRITE_AFTER_END or double-end the response.
-  const { send, safeEnd } = openSseStream(res);
-  // Server-console visibility for the (multi-GB) local pip install — the SSE
-  // stream otherwise surfaces progress only in the browser.
-  const installLog = createInstallLogger({ installer: 'Image Gen packages', target: parsed.data.pythonPath });
-  const emit = (ev) => { installLog.onEvent(ev); send(ev); };
-  installLog.start();
-  const { promise, kill } = installPackages(parsed.data.pythonPath, parsed.data.packages, emit);
-  promise.then(() => {
-    // Drop the now-stale setup-check snapshot before the client re-runs the
-    // probe on `complete` — without this it would read the pre-install
-    // missing-packages list back from cache.
-    invalidateSetupCheck(parsed.data.pythonPath);
-    safeEnd();
-  });
-
-  // Client navigation away should kill pip — a torch upgrade can run for
-  // 10+ minutes and would otherwise keep going invisibly.
-  req.on('close', () => { installLog.cancel(); kill(); safeEnd(); });
-});
+// Local-mode setup automation (python probe/venv, FLUX.2 install/status,
+// HF token store, pip installer) lives in its own router.
+router.use("/setup", setupRouter);
 
 export default router;
